+++ NEWS.md
@@ -1,5 +1,15 @@
 # ncdfCF (development version)
 
+* NetCDF groups and user-defined data types are now supported. Information on
+the structure of UDTs can now be examined. This is effectively only supported
+for the "compound" sub-type, for scalar values only.
+* Better support for discrete axis dimensions, with a separate S4 class.
+* Data is read into the most compact form possible. This saves a significant 
+amount of memory when large integer variables are read as they would remain
+integers rather than the default numeric type.
+* Improvements in printing object details.
+* GHA enabled
+
 # ncdfCF 0.1.1
 
 * `objects_by_standard_name()` will list objects in the NetCDF resource that